+++ submit_ce/implementations/compile/compile_at_gcp_service.py
@@ -8,8 +8,7 @@ from submit_ce.api import User, Client, Submission
 from submit_ce.api.CompileService import CompileService
 from submit_ce.api.domain.event.process import Result
 from submit_ce.api.domain.process import ProcessStatus
-import sys
-sys.path.append('submit_ce/implementations/compile')
+
 from submit_ce.implementations.compile.compile_at_gcp import PreflightOption, DEFAULT_MAX_APPEND_FILES, DEFAULT_MAX_TEX_FILES, \
     DEFAULT_COMPILATION_TIMEOUT, compile_submission
 
@@ -82,3 +81,4 @@ class GcpCompileAtLegacy(CompileService):
 
     def check(self, process_id: str, user: User, client: Client) -> ProcessStatus:
         pass
+